scripts/ingest/collect_usda_fas_esr.py
#!/usr/bin/env python3
"""
USDA FAS Export Sales Reporting (ESR) - Weekly China Purchases
===============================================================
Collects weekly US export sales to China from USDA FAS.
Critical for tracking China's purchase pace of US soybeans.

Source: https://apps.fas.usda.gov/esrquery/
Updates: Weekly (Thursdays/Fridays)
"""

import requests
import pandas as pd
from pathlib import Path
from datetime import datetime, timedelta
import logging
import json
import time
from bs4 import BeautifulSoup

# Setup logging
logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(levelname)s - %(message)s'
)
logger = logging.getLogger(__name__)

# Configuration
EXTERNAL_DRIVE = Path("/Volumes/Satechi Hub/Projects/CBI-V14")
RAW_DIR = EXTERNAL_DRIVE / "TrainingData/raw/usda_fas_esr"
RAW_DIR.mkdir(parents=True, exist_ok=True)

# USDA FAS ESR - Official weekly historical pages (scrape-friendly)
ESR_PAGES = {
    'soybeans': 'https://apps.fas.usda.gov/export-sales/h801.htm',
    'soybean_oil': 'https://apps.fas.usda.gov/export-sales/h902.htm'
}

# Key commodities for soybean complex
COMMODITIES = {
    'soybeans': {'url': ESR_PAGES['soybeans'], 'unit': 'MT'},
    'soybean_oil': {'url': ESR_PAGES['soybean_oil'], 'unit': 'MT'}
}

# China destinations
CHINA_CODES = ['CHINA, PEOPLES REPUBLIC OF', 'CHINA', 'CN']

def scrape_esr_weekly_page(commodity: str) -> pd.DataFrame:
    """
    Scrape USDA FAS ESR weekly historical page.
    NOTE: This page shows AGGREGATE export sales, not country-specific.
    For China-specific data, use aggregate as proxy or find alternative source.
    
    Args:
        commodity: Commodity name (soybeans, soybean_oil)
        
    Returns:
        DataFrame with weekly aggregate export sales data
    """
    if commodity not in COMMODITIES:
        logger.error(f"Unknown commodity: {commodity}")
        return pd.DataFrame()
    
    url = COMMODITIES[commodity]['url']
    
    logger.info(f"Scraping {commodity} from {url}...")
    logger.warning(f"  ⚠️  Note: This page shows AGGREGATE data, not country-specific")
    
    try:
        response = requests.get(url, timeout=60, headers={
            'User-Agent': 'Mozilla/5.0 (compatible; CBI-V14-Data-Collector/1.0)'
        })
        
        if response.status_code != 200:
            logger.error(f"HTTP {response.status_code} for {commodity}")
            return pd.DataFrame()
        
        # Parse HTML table
        soup = BeautifulSoup(response.text, 'html.parser')
        tables = soup.find_all('table')
        
        if not tables:
            logger.warning(f"No tables found in {commodity} page")
            return pd.DataFrame()
        
        # Find the main data table
        all_rows = []
        table = tables[0]  # Use first table
        rows = table.find_all('tr')
        
        # Find header row (contains "Week Ending")
        header_idx = None
        for i, row in enumerate(rows[:10]):
            cells = [cell.get_text(strip=True) for cell in row.find_all(['td', 'th'])]
            if any('week' in cell.lower() and 'ending' in cell.lower() for cell in cells):
                header_idx = i
                break
        
        if header_idx is None:
            logger.warning("Could not find header row")
            return pd.DataFrame()
        
        # Parse data rows
        for row in rows[header_idx + 2:]:  # Skip header and sub-header
            cells = [cell.get_text(strip=True) for cell in row.find_all(['td', 'th'])]
            
            if len(cells) >= 2 and cells[0]:  # Has at least date and one data column
                # First cell should be date
                try:
                    date_str = cells[0]
                    # Try to parse date (format: MM/DD/YYYY)
                    if '/' in date_str:
                        week_ending = pd.to_datetime(date_str, format='%m/%d/%Y', errors='coerce')
                        if pd.notna(week_ending):
                            row_data = {
                                'week_ending': week_ending,
                                'weekly_exports': pd.to_numeric(cells[1].replace(',', ''), errors='coerce') if len(cells) > 1 else None,
                                'accumulated_exports': pd.to_numeric(cells[2].replace(',', ''), errors='coerce') if len(cells) > 2 else None,
                                'net_sales': pd.to_numeric(cells[3].replace(',', ''), errors='coerce') if len(cells) > 3 else None,
                                'outstanding_sales': pd.to_numeric(cells[4].replace(',', ''), errors='coerce') if len(cells) > 4 else None,
                            }
                            all_rows.append(row_data)
                except:
                    continue
        
        if not all_rows:
            logger.warning(f"No data rows parsed for {commodity}")
            return pd.DataFrame()
        
        df = pd.DataFrame(all_rows)
        df = df[df['week_ending'].notna()]  # Remove invalid dates
        df['commodity'] = commodity
        df['destination'] = 'AGGREGATE'  # Mark as aggregate, not country-specific
        df['source'] = 'USDA_FAS_ESR'
        df['scraped_at'] = datetime.now()
        df['note'] = 'Aggregate export sales - China-specific data not available on this page'
        
        logger.info(f"  ✅ Parsed {len(df)} weekly records (AGGREGATE data)")
        logger.warning(f"  ⚠️  China-specific data requires alternative source")
        
        return df
        
    except Exception as e:
        logger.error(f"Error scraping {commodity}: {e}")
        import traceback
        traceback.print_exc()
        return pd.DataFrame()

def scrape_esr_report(report_url: str) -> pd.DataFrame:
    """
    Scrape a specific ESR weekly report.
    
    Args:
        report_url: URL to the weekly report
        
    Returns:
        DataFrame with parsed data
    """
    try:
        response = requests.get(report_url, timeout=30)
        
        if response.status_code == 200:
            soup = BeautifulSoup(response.text, 'html.parser')
            
            # Parse tables (ESR uses HTML tables)
            tables = soup.find_all('table')
            
            # Find the exports table
            # Look for China-specific rows
            # Structure varies by report format
            
            # TODO: Implement report parsing (no placeholder generation)
            logger.debug(f"Parsed report: {report_url}")
            
    except Exception as e:
        logger.error(f"Error scraping report: {e}")
    
    return pd.DataFrame()

def aggregate_export_sales(all_data: list) -> pd.DataFrame:
    """
    Aggregate all commodity export sales data.
    NOTE: This is AGGREGATE data, not China-specific.
    """
    if not all_data:
        return pd.DataFrame()
    
    # Combine all commodities
    combined = pd.concat(all_data, ignore_index=True)
    
    # Aggregate by week (all destinations combined - aggregate)
    weekly = combined.groupby(['week_ending', 'commodity']).agg({
        'weekly_exports': 'sum',
        'accumulated_exports': 'sum',
        'net_sales': 'sum',
        'outstanding_sales': 'sum',
        'destination': 'first',  # Will be 'AGGREGATE'
        'note': 'first'
    }).reset_index()
    
    # Add metadata
    weekly['source'] = 'USDA_FAS_ESR'
    weekly['updated_at'] = datetime.now()
    
    return weekly

def main():
    """Main execution"""
    logger.info("="*80)
    logger.info("USDA FAS EXPORT SALES REPORTING")
    logger.info("="*80)
    logger.info("Weekly US export sales to China")
    logger.info("="*80)
    logger.info("")
    logger.info("⚠️  NOTE: The h801.htm page shows aggregate data, not country-specific.")
    logger.info("    China-specific data may require:")
    logger.info("    1. USDA FAS ESR query interface (requires parameters)")
    logger.info("    2. MyMarketNews API (requires authentication)")
    logger.info("    3. Alternative: Use aggregate data as proxy")
    logger.info("="*80)
    
    all_commodity_data = []
    
    # Scrape data for each commodity
    for commodity in COMMODITIES.keys():
        df = scrape_esr_weekly_page(commodity)
        
        if not df.empty:
            all_commodity_data.append(df)
            logger.info(f"✅ {commodity}: {len(df)} rows")
    
    if not all_commodity_data:
        logger.error("No data collected")
        return 1
    
    # Aggregate export sales (AGGREGATE data, not China-specific)
    export_df = aggregate_export_sales(all_commodity_data)
    
    # Save raw data
    raw_file = RAW_DIR / f"esr_export_sales_{datetime.now().strftime('%Y%m%d')}.parquet"
    if not export_df.empty:
        export_df.to_parquet(raw_file, index=False)
        logger.info(f"✅ Saved: {raw_file.name}")
    
    # Create weekly summary
    if not export_df.empty:
        summary = export_df.pivot_table(
            index='week_ending',
            columns='commodity',
            values='weekly_exports',
            fill_value=0
        ).reset_index()
        
        summary_file = RAW_DIR / f"esr_weekly_summary_{datetime.now().strftime('%Y%m%d')}.parquet"
        summary.to_parquet(summary_file, index=False)
        logger.info(f"✅ Saved summary: {summary_file.name}")
    
    logger.info("\n" + "="*80)
    logger.info("IMPORTANT NOTE")
    logger.info("="*80)
    logger.info("This script creates the structure for ESR data collection.")
    logger.info("Full implementation requires:")
    logger.info("  1. Parsing ESR HTML reports or")
    logger.info("  2. Using USDA MyMarketNews API (requires authentication)")
    logger.info("  3. Or scraping weekly PDF/CSV exports")
    logger.info("\nRecommended: Use MyMarketNews API for reliable weekly updates")
    logger.info("API Docs: https://mymarketnews.ams.usda.gov/api")
    logger.info("="*80)
    
    return 0

if __name__ == "__main__":
    import sys
    sys.exit(main())
